diagnose firewall inter-cluster: add support for LB mode
When I checked the firewall inter-cluster with LB mode I noticed that:

A. The encapsulated traffic (port 4500) received in some platforms
   on LB nodeport
B. I'm not sure if it's specific to LB mode,sometimes the client pod
finished running before the tcpdump server started running.

This PR updates the server tcpdump pod to listen also on nodeport in case
SM is deployed in LB mode and address 'B' by adding short delay
to client pod.

Fixes: https://github.com/submariner-io/submariner/issues/1849

Signed-off-by: yboaron <yboaron@redhat.com>